Use Android SCAN_RESULTS_AVAILABLE_ACTION
This is almost a copy from https://github.com/jeffcharles/WifiWizard/tree/scan-and-list-master
And, no, I'm quite unhappy with the state of the code still. It has a lot of duplication
(why is there a `scan` and `startScan` method?) and the `scan` implementation
has some weird timeout thread going on which might or might not be necessary.